added fmt and sentry dsn variable
