+++ server.go
@@ -3,21 +3,19 @@ package main
 import (
 	"encoding/json"
 	"fmt"
+	"github.com/getsentry/sentry-go"
 	"github.com/gorilla/mux"
+	"github.com/jasonlvhit/gocron"
+	"gorm.io/driver/postgres"
+	"gorm.io/gorm"
 	"io/ioutil"
 	"log"
 	"net/http"
 	"os"
 	dbops "sky-meter/dbops"
 	httpreponser "sky-meter/httpres"
-	"github.com/jasonlvhit/gocron"
-	"github.com/getsentry/sentry-go"
-	  "gorm.io/driver/postgres"
-  "gorm.io/gorm"
-  models "sky-meter/models"
+	models "sky-meter/models"
 )
-
-
 
 func homeLink(w http.ResponseWriter, r *http.Request) {
 	fmt.Fprintf(w, "Welcome to sky-meter")
@@ -30,22 +28,20 @@ func getStats(w http.ResponseWriter, r *http.Request) {
 	return
 }
 
-
-
 func httpSyntheticCheck(endpoint string, time uint64) {
-  gocron.Every(time).Second().Do(callEndpoint, endpoint)
-  <-gocron.Start()
-  fmt.Println(time)
+	gocron.Every(time).Second().Do(callEndpoint, endpoint)
+	<-gocron.Start()
 }
 
 func callEndpoint(endpoint string) {
 	httpresdata, _ := httpreponser.GetHttpdata(endpoint)
-	fmt.Println(string(httpresdata))
+	log.Println(string(httpresdata))
 }
 
 func main() {
+	senenv := os.Getenv("sentry_dsn")
 	senterr := sentry.Init(sentry.ClientOptions{
-		Dsn: "addYourDSNHere",
+		Dsn: senenv,
 		// Set TracesSampleRate to 1.0 to capture 100%
 		// of transactions for performance monitoring.
 		// We recommend adjusting this value in production,
@@ -54,7 +50,6 @@ func main() {
 	if senterr != nil {
 		log.Fatalf("sentry.Init: %s", senterr)
 	}
-
 
 	jsonFile, err := os.Open("input.json")
 	if err != nil {
@@ -66,26 +61,24 @@ func main() {
 
 	var endpoints models.JsonInput
 
-	json.Unmarshal(byteValue, &endpoints)	
+	json.Unmarshal(byteValue, &endpoints)
 
-db, err := gorm.Open(postgres.New(postgres.Config{
-    DSN:                  "host=localhost user=postgres password=postgres dbname=postgres port=5433 sslmode=disable TimeZone=Asia/Shanghai",
-    PreferSimpleProtocol: true, // disables implicit prepared statement usage
-  }), &gorm.Config{})
+	db, err := gorm.Open(postgres.New(postgres.Config{
+		DSN:                  "host=localhost user=postgres password=postgres dbname=postgres port=5433 sslmode=disable TimeZone=Asia/Shanghai",
+		PreferSimpleProtocol: true, // disables implicit prepared statement usage
+	}), &gorm.Config{})
 
-  if err != nil {
-    log.Println(err)
-  }
+	if err != nil {
+		log.Println(err)
+	}
 
-  dbops.InitialMigration(db)
-  dbops.InsertUrlsToDb(db,endpoints)
+	dbops.InitialMigration(db)
+	dbops.InsertUrlsToDb(db, endpoints)
 
-	fmt.Println("listening on port 8080")
+	log.Println("listening on port 8080")
 	router := mux.NewRouter().StrictSlash(true)
 	router.HandleFunc("/", homeLink)
 	router.HandleFunc("/stats", getStats).Methods("GET")
 	log.Fatal(http.ListenAndServe(":8080", router))
-
-
 
 }
